+++ afanasy/trunk/src/libafanasy/taskprogress.h
@@ -19,12 +19,12 @@ public:
 
    uint32_t state;         ///< state per block per task.
    int8_t   percent;       ///< percent per block per task.
-   int32_t  frame;         ///< frame per block per task.
+   int64_t  frame;         ///< frame per block per task.
    int8_t   percentframe;  ///< frame percent per block per task.
    int32_t  starts_count;  ///< number of starts per block per task.
    int32_t  errors_count;  ///< Number of times task finished with errors .
-   uint32_t time_start;    ///< start time.
-   uint32_t time_done;     ///< Task finish time ( or last update time if still running ).
+   int64_t time_start;     ///< start time.
+   int64_t time_done;      ///< Task finish time ( or last update time if still running ).
 
    std::string hostname;           ///< Host, last event occurs where.
